Fix issue with audio player stop not responding
BufferedDataSource waits indefinitely for more data if there is
no more data available causing the player's message handler to be
blocked. If a stop is called during this time the message handler
isn't available to service the stop message and therefore doesn't
handle the event. This can happen when, for example, the client
stop writing audio data and then after couple of seconds calls the
stop method.